--- a/causallearn/utils/Fas.py
+++ b/causallearn/utils/Fas.py
@@ -4,6 +4,7 @@ from causallearn.graph.GeneralGraph import GeneralGraph
 from causallearn.graph.Edges import Edges
 from causallearn.utils.ChoiceGenerator import ChoiceGenerator
 from copy import deepcopy
+from causallearn.search.ConstraintBased.FCI import citest_cache
 
 
 def possible_parents(node_x, adjx, knowledge=None):
@@ -47,7 +48,12 @@ def searchAtDepth0(data, nodes, adjacencies, sep_sets, independence_test_method=
             print(nodes[i + 1].get_name())
 
         for j in range(i+1, len(nodes)):
-            p_value = independence_test_method(data, i, j, tuple(empty))
+            ijS_key = (i, j, frozenset())
+            if ijS_key in citest_cache:
+                p_value = citest_cache[ijS_key]
+            else:
+                p_value = independence_test_method(data, i, j, tuple(empty))
+                citest_cache[ijS_key] = p_value
             independent = p_value > alpha
             no_edge_required = True if knowledge is None else \
                 ((not knowledge.is_required(nodes[i], nodes[j])) or knowledge.is_required(nodes[j], nodes[i]))
@@ -80,7 +86,15 @@ def searchAtDepth(data, depth, nodes, adjacencies, sep_sets, independence_test_m
                     cond_set = [nodes.index(ppx[index]) for index in choice]
                     choice = cg.next()
 
-                    p_value = independence_test_method(data, i, nodes.index(adjx[j]), tuple(cond_set))
+                    Y = nodes.index(adjx[j])
+                    X, Y = (i, Y) if (i < Y) else (Y, i)
+                    XYS_key = (X, Y, frozenset(cond_set))
+                    if XYS_key in citest_cache:
+                        p_value = citest_cache[XYS_key]
+                    else:
+                        p_value = independence_test_method(data, X, Y, tuple(cond_set))
+                        citest_cache[XYS_key] = p_value
+
                     independent = p_value > alpha
 
                     no_edge_required = True if knowledge is None else \