--- a/causallearn/utils/PCUtils/SkeletonDiscovery.py
+++ b/causallearn/utils/PCUtils/SkeletonDiscovery.py
@@ -4,9 +4,10 @@ import numpy as np
 
 from causallearn.graph.GraphClass import CausalGraph
 from causallearn.utils.PCUtils.Helper import append_value
+from causallearn.utils.cit import chisq, gsq
 
 
-def skeleton_discovery(data, alpha, indep_test, stable=True, background_knowledge=None):
+def skeleton_discovery(data, alpha, indep_test, stable=True, background_knowledge=None, verbose=False):
     '''
     Perform skeleton discovery
 
@@ -31,8 +32,21 @@ def skeleton_discovery(data, alpha, indep_test, stable=True, background_knowledg
 
     no_of_var = data.shape[1]
     cg = CausalGraph(no_of_var)
-    cg.data = data
     cg.set_ind_test(indep_test)
+    if indep_test == chisq or indep_test == gsq:
+        # if dealing with discrete data, data is numpy.ndarray with n rows m columns,
+        # for each column, translate the discrete values to int indexs starting from 0,
+        #   e.g. [45, 45, 6, 7, 6, 7] -> [2, 2, 0, 1, 0, 1]
+        #        ['apple', 'apple', 'pear', 'peach', 'pear'] -> [0, 0, 2, 1, 2]
+        # in old code, its presumed that discrete `data` is already indexed,
+        # but here we make sure it's in indexed form, so allow more user input e.g. 'apple' ..
+        def _unique(column):
+            return np.unique(column, return_inverse=True)[1]
+        cg.is_discrete = True
+        cg.data = np.apply_along_axis(_unique, 0, data).astype(np.int64)
+        cg.cardinalities = np.max(cg.data, axis=0) + 1
+    else:
+        cg.data = data
 
     depth = -1
     while cg.max_degree() - 1 > depth:
@@ -48,7 +62,7 @@ def skeleton_discovery(data, alpha, indep_test, stable=True, background_knowledg
                     if background_knowledge is not None and (
                             background_knowledge.is_forbidden(cg.G.nodes[x], cg.G.nodes[y])
                             and background_knowledge.is_forbidden(cg.G.nodes[y], cg.G.nodes[x])):
-                        print('%d ind %d | %s with background background_knowledge\n' % (x, y, S))
+                        if verbose: print('%d ind %d | %s with background background_knowledge\n' % (x, y, S))
                         if not stable:
                             edge1 = cg.G.get_edge(cg.G.nodes[x], cg.G.nodes[y])
                             if edge1 is not None:
@@ -65,7 +79,7 @@ def skeleton_discovery(data, alpha, indep_test, stable=True, background_knowledg
                     else:
                         p = cg.ci_test(x, y, S)
                         if p > alpha:
-                            print('%d ind %d | %s with p-value %f\n' % (x, y, S, p))
+                            if verbose: print('%d ind %d | %s with p-value %f\n' % (x, y, S, p))
                             if not stable:
                                 edge1 = cg.G.get_edge(cg.G.nodes[x], cg.G.nodes[y])
                                 if edge1 is not None:
@@ -80,7 +94,7 @@ def skeleton_discovery(data, alpha, indep_test, stable=True, background_knowledg
                                 append_value(cg.sepset, y, x, S)
                             break
                         else:
-                            print('%d dep %d | %s with p-value %f\n' % (x, y, S, p))
+                            if verbose: print('%d dep %d | %s with p-value %f\n' % (x, y, S, p))
 
         for (x, y) in list(set(edge_removal)):
             edge1 = cg.G.get_edge(cg.G.nodes[x], cg.G.nodes[y])