… tests/TestData/bnlearn_discrete_10000/alarm.txt
… tests/TestData/bnlearn_discrete_10000/asia.txt
… tests/TestData/bnlearn_discrete_10000/barley.txt
… tests/TestData/bnlearn_discrete_10000/cancer.txt
… tests/TestData/bnlearn_discrete_10000/child.txt
… tests/TestData/bnlearn_discrete_10000/earthquake.txt
… tests/TestData/bnlearn_discrete_10000/hailfinder.txt
… tests/TestData/bnlearn_discrete_10000/hepar2.txt
… tests/TestData/bnlearn_discrete_10000/insurance.txt
… tests/TestData/bnlearn_discrete_10000/sachs.txt
… tests/TestData/bnlearn_discrete_10000/survey.txt
… tests/TestData/bnlearn_discrete_10000/water.txt
… tests/TestData/bnlearn_discrete_10000/win95pts.txt